Day 1 : Doubly Linked List Change
Added a new Java file to perform the Insertion, Deletion and searching
operation on Doubly Linked List

Insertion - for Insertion I have created two methods, one is to insert a
node at the end of the list and another is to insert a node at the front
of the list.

Deletion - For deletion I have handled three Cases : If node which is to
be deleted is first Node, If node to be deleted if middle node and else
if it is a last node.

Search - In search method I am finding the node and returning it's
index. And if that node is not present then I am printing the message
that this node is not present.